feat: add @event and improve test coverage
This commit adds an @event decorator so that function developers can
explicitly specify the type and source of returned events. Any function
using the @event decorator will have its return value set to the data
property of a CloudEvent, created by the server and returned to the
caller.

I had to split the tests up, because I could not get Python to load
a 'func' module twice from two different locations, with different
implementations.

Signed-off-by: Lance Ball <lball@redhat.com>